refactor delete files
