+++ napari_allencell_annotator/view/images_view.py
@@ -105,7 +105,7 @@ class ImagesView(QFrame):
         self.file_widget.files_added.connect(self._toggle_shuffle)
 
         self.shuffle.toggled.connect(self._update_shuff_text)
-        self.delete.clicked.connect(self._delete_clicked)
+        # self.delete.clicked.connect(self._delete_clicked)
         self.file_widget.currentItemChanged.connect(self._display_img)
 
     def _update_shuff_text(self, checked: bool):
@@ -133,23 +133,23 @@ class ImagesView(QFrame):
         self._toggle_shuffle(False)
         self.toggle_add(True)
 
-    def _delete_clicked(self):
-        """Ask user to approve a list of files to delete from the file list."""
-        if len(self.file_widget.checked) > 0:
-            msg: str = "Delete these files from the list?"
-            names: Set[str] = set()
-            for item in self.file_widget.checked:
-                names.add("--- " + item.file_path)
-            msg_box = ScrollablePopup(msg, names)
-            if msg_box.exec() == QDialog.Accepted:
-                self.file_widget.delete_checked()
-        else:
-            proceed: bool = Popup.make_popup("Remove all images?")
-            if proceed:
-                self.file_widget.clear_all()
-                self.reset_buttons()
-
-        self.update_num_files_label(self.file_widget.count())
+    # def _delete_clicked(self):
+    #     """Ask user to approve a list of files to delete from the file list."""
+    #     if len(self.file_widget.checked) > 0:
+    #         msg: str = "Delete these files from the list?"
+    #         names: Set[str] = set()
+    #         for item in self.file_widget.checked:
+    #             names.add("--- " + item.file_path)
+    #         msg_box = ScrollablePopup(msg, names)
+    #         if msg_box.exec() == QDialog.Accepted:
+    #             self.file_widget.delete_checked()
+    #     else:
+    #         proceed: bool = Popup.make_popup("Remove all images?")
+    #         if proceed:
+    #             self.file_widget.clear_all()
+    #             self.reset_buttons()
+    #
+    #     self.update_num_files_label(self.file_widget.count())
 
     def alert(self, alert_msg: str):
         """
@@ -220,4 +220,4 @@ class ImagesView(QFrame):
 
     def update_num_files_label(self, num_files: int) -> None:
         """Update num_files_label to show the current number of image files"""
-        self.num_files_label.setText(f"Loaded files: {num_files}")
+        self.num_files_label.setText(f"Image files: {num_files}")
